cohorts/2023/week_7_project/requirements.txt
requests==2.28.1
pandas==1.4.4
pandas-gbq==0.19.1
pathlib==1.0.1
pyarrow==11.0.0
prefect==2.8.4
python-decouple==3.8
prefect-gcp[cloud_storage]==0.2.4
prefect-dbt[bigquery]==0.3.1
